@@ -1,15 +1,15 @@
 // Code generated by go-bindata. DO NOT EDIT.
 // sources:
-// ../datasource/migrations/1555589258_create_users.down.sql (19B)
-// ../datasource/migrations/1555589258_create_users.up.sql (272B)
-// ../datasource/migrations/1555589937_create_domains.down.sql (21B)
-// ../datasource/migrations/1555589937_create_domains.up.sql (164B)
-// ../datasource/migrations/1585465452_create_domain_users.down.sql (25B)
-// ../datasource/migrations/1585465452_create_domain_users.up.sql (323B)
-// ../datasource/migrations/1643446918_add_email_users.down.sql (31B)
-// ../datasource/migrations/1643446918_add_email_users.up.sql (340B)
-// ../datasource/migrations/1732505546_add_email_aliases.down.sql (32B)
-// ../datasource/migrations/1732505546_add_email_aliases.up.sql (260B)
+// 1555589258_create_users.down.sql (19B)
+// 1555589258_create_users.up.sql (272B)
+// 1555589937_create_domains.down.sql (21B)
+// 1555589937_create_domains.up.sql (164B)
+// 1585465452_create_domain_users.down.sql (25B)
+// 1585465452_create_domain_users.up.sql (323B)
+// 1643446918_add_email_users.down.sql (31B)
+// 1643446918_add_email_users.up.sql (340B)
+// 1732505546_add_email_aliases.down.sql (32B)
+// 1732505546_add_email_aliases.up.sql (260B)
 
 package migrationData
 
@@ -19,7 +19,6 @@ import (
 	"crypto/sha256"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"strings"
@@ -29,7 +28,7 @@ import (
 func bindataRead(data []byte, name string) ([]byte, error) {
 	gz, err := gzip.NewReader(bytes.NewBuffer(data))
 	if err != nil {
-		return nil, fmt.Errorf("read %q: %v", name, err)
+		return nil, fmt.Errorf("read %q: %w", name, err)
 	}
 
 	var buf bytes.Buffer
@@ -37,7 +36,7 @@ func bindataRead(data []byte, name string) ([]byte, error) {
 	clErr := gz.Close()
 
 	if err != nil {
-		return nil, fmt.Errorf("read %q: %v", name, err)
+		return nil, fmt.Errorf("read %q: %w", name, err)
 	}
 	if clErr != nil {
 		return nil, err
@@ -381,15 +380,20 @@ var _bindata = map[string]func() (*asset, error){
 	"1732505546_add_email_aliases.up.sql":     _1732505546_add_email_aliasesUpSql,
 }
 
+// AssetDebug is true if the assets were built with the debug flag enabled.
+const AssetDebug = false
+
 // AssetDir returns the file names below a certain
 // directory embedded in the file by go-bindata.
 // For example if you run go-bindata on data/... and data contains the
 // following hierarchy:
-//     data/
-//       foo.txt
-//       img/
-//         a.png
-//         b.png
+//
+//	data/
+//	  foo.txt
+//	  img/
+//	    a.png
+//	    b.png
+//
 // then AssetDir("data") would return []string{"foo.txt", "img"},
 // AssetDir("data/img") would return []string{"a.png", "b.png"},
 // AssetDir("foo.txt") and AssetDir("notexist") would return an error, and
@@ -422,16 +426,16 @@ type bintree struct {
 }
 
 var _bintree = &bintree{nil, map[string]*bintree{
-	"1555589258_create_users.down.sql":        &bintree{_1555589258_create_usersDownSql, map[string]*bintree{}},
-	"1555589258_create_users.up.sql":          &bintree{_1555589258_create_usersUpSql, map[string]*bintree{}},
-	"1555589937_create_domains.down.sql":      &bintree{_1555589937_create_domainsDownSql, map[string]*bintree{}},
-	"1555589937_create_domains.up.sql":        &bintree{_1555589937_create_domainsUpSql, map[string]*bintree{}},
-	"1585465452_create_domain_users.down.sql": &bintree{_1585465452_create_domain_usersDownSql, map[string]*bintree{}},
-	"1585465452_create_domain_users.up.sql":   &bintree{_1585465452_create_domain_usersUpSql, map[string]*bintree{}},
-	"1643446918_add_email_users.down.sql":     &bintree{_1643446918_add_email_usersDownSql, map[string]*bintree{}},
-	"1643446918_add_email_users.up.sql":       &bintree{_1643446918_add_email_usersUpSql, map[string]*bintree{}},
-	"1732505546_add_email_aliases.down.sql":   &bintree{_1732505546_add_email_aliasesDownSql, map[string]*bintree{}},
-	"1732505546_add_email_aliases.up.sql":     &bintree{_1732505546_add_email_aliasesUpSql, map[string]*bintree{}},
+	"1555589258_create_users.down.sql":        {_1555589258_create_usersDownSql, map[string]*bintree{}},
+	"1555589258_create_users.up.sql":          {_1555589258_create_usersUpSql, map[string]*bintree{}},
+	"1555589937_create_domains.down.sql":      {_1555589937_create_domainsDownSql, map[string]*bintree{}},
+	"1555589937_create_domains.up.sql":        {_1555589937_create_domainsUpSql, map[string]*bintree{}},
+	"1585465452_create_domain_users.down.sql": {_1585465452_create_domain_usersDownSql, map[string]*bintree{}},
+	"1585465452_create_domain_users.up.sql":   {_1585465452_create_domain_usersUpSql, map[string]*bintree{}},
+	"1643446918_add_email_users.down.sql":     {_1643446918_add_email_usersDownSql, map[string]*bintree{}},
+	"1643446918_add_email_users.up.sql":       {_1643446918_add_email_usersUpSql, map[string]*bintree{}},
+	"1732505546_add_email_aliases.down.sql":   {_1732505546_add_email_aliasesDownSql, map[string]*bintree{}},
+	"1732505546_add_email_aliases.up.sql":     {_1732505546_add_email_aliasesUpSql, map[string]*bintree{}},
 }}
 
 // RestoreAsset restores an asset under the given directory.
@@ -448,7 +452,7 @@ func RestoreAsset(dir, name string) error {
 	if err != nil {
 		return err
 	}
-	err = ioutil.WriteFile(_filePath(dir, name), data, info.Mode())
+	err = os.WriteFile(_filePath(dir, name), data, info.Mode())
 	if err != nil {
 		return err
 	}
